+++ week02/mission02/src/App.tsx
@@ -1,13 +1,13 @@
-import './App.css'
+import "./App.css";
 import ContextPage from "./context/ContextPage";
-import { ThemeProvider } from './context/ThemeProvider';
+import { ThemeProvider } from "./context/ThemeProvider";
 
 export function App() {
   return (
     <ThemeProvider>
       <ContextPage></ContextPage>
     </ThemeProvider>
-  )
+  );
 }
 
-export default App
+export default App;
