--- week02/week02-1/src/App.tsx
@@ -1,14 +1,0 @@
-import './App.css'
-import Todo from '../components/Todo'
-import { TodoProvider } from '../src/context/TodoContext'
-
-function App() {
-
-  return (
-    <TodoProvider>
-      <Todo></Todo>
-    </TodoProvider>
-  )
-}
-
-export default App
